Revert "commit6:longer float input"
This reverts commit 402fba3d2a10555d60d2b31e61cef61dd9051c1c.

+++ project1.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-    double a = 0, b = 0;
-    scanf("%lf %lf", &a, &b);
+    float a = 0, b = 0;
+    scanf("%f %f", &a, &b);
     printf("%lld", a * b);
 }
